// src/app/models/trip-model.ts
export class Trip{
  tripTitle: string;
  destination: string;
  startDate: string;
  endDate: string;
  price: number;
  places: number;
  description: string;
  url: string;
  reservedPlaces: number;
  sumOfRatings: number;
  key: string;
  rate: number;
  numberOfRatings?: number;
}